refactor(torghut): split archive persistence types
